backend/requirements.mvp.txt
# 🚀 MVP精简版依赖 - 只保留核心功能
# 目标: 镜像大小 < 2GB (从8.5GB大幅精简)

# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database (保留)
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Cache
redis==5.2.1

# Auth (保留)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Validation
pydantic==2.10.4
pydantic-settings==2.7.1

# Market Data (保留核心)
httpx==0.27.2
yfinance==0.2.50

# News Data (简化: 只保留轻量级爬虫)
aiohttp==3.11.11
beautifulsoup4==4.12.3
feedparser==6.0.11
requests>=2.28.0

# AI - MVP精简版 (用OpenAI API代替本地模型)
openai==1.55.0
pinecone-client==5.0.1

# Data Processing (简化)
numpy==1.26.4
pandas==2.2.3

# Task Queue (保留)
celery[redis]==5.4.0
apscheduler==3.11.0

# Utils
python-dotenv==1.0.1
aiofiles==24.1.0
